LPS-162363 Split custom hooks logic
